+++ bindings/java/examples/basic-app/src/test/java/org/example/LibraryTest.java
@@ -1,0 +1,14 @@
+package org.example;
+
+import org.junit.Test;
+
+public class LibraryTest {
+    @Test
+    public void testSomeLibraryMethod() {
+        try {
+            ExampleApp.nodeInfo();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+}