+++ bindings/java/native/src/classes/client/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod balance;
+pub mod client_builder;
+pub mod full_node_api;
+pub mod mqtt;